pyproject.toml
[project]
name = "airflow-learning"
version = "0.1.0"
dependencies = [
    "apache-airflow==3.0.6",
    "apache-airflow-providers-postgres==6.3.0",
]
